refactor: Move the vela api to the apis folder
